hw3: config properties fix
