+++ bot/src/main/java/edu/java/bot/BotApplication.java
@@ -1,16 +1,11 @@
 package edu.java.bot;
 
-import edu.java.bot.configuration.ApplicationConfig;
-import edu.java.bot.configuration.ScrapperConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 
 @SpringBootApplication
-@EnableConfigurationProperties({
-        ApplicationConfig.class,
-        ScrapperConfig.class
-})
+@ConfigurationPropertiesScan
 public class BotApplication {
     public static void main(String[] args) {
         SpringApplication.run(BotApplication.class, args);
