+++ bot/src/main/java/edu/java/bot/configuration/BotConfiguration.java
@@ -1,6 +1,7 @@
 package edu.java.bot.configuration;
 
 import com.pengrad.telegrambot.TelegramBot;
+import edu.java.bot.configuration.properties.ApplicationConfig;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
