+++ bot/src/main/java/edu/java/bot/configuration/properties/ApplicationConfig.java
@@ -1,13 +1,13 @@
-package edu.java.configuration;
+package edu.java.bot.configuration.properties;
 
 import jakarta.validation.constraints.NotEmpty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.validation.annotation.Validated;
 
 @Validated
-@ConfigurationProperties(prefix = "bot-client", ignoreUnknownFields = false)
-public record BotConfig(
-        @NotEmpty
-        String baseUrl
+@ConfigurationProperties(prefix = "app", ignoreUnknownFields = false)
+public record ApplicationConfig(
+    @NotEmpty
+    String telegramToken
 ) {
 }
