+++ bot/src/main/java/edu/java/bot/configuration/properties/ScrapperConfig.java
@@ -1,13 +1,13 @@
-package edu.java.bot.configuration;
+package edu.java.bot.configuration.properties;
 
 import jakarta.validation.constraints.NotEmpty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.validation.annotation.Validated;
 
 @Validated
-@ConfigurationProperties(prefix = "app", ignoreUnknownFields = false)
-public record ApplicationConfig(
-    @NotEmpty
-    String telegramToken
+@ConfigurationProperties(prefix = "scrapper-client", ignoreUnknownFields = false)
+public record ScrapperConfig(
+        @NotEmpty
+        String baseUrl
 ) {
 }
