+++ scrapper/src/main/java/edu/java/ScrapperApplication.java
@@ -1,22 +1,18 @@
 package edu.java;
 
-import edu.java.configuration.ApplicationConfig;
-import edu.java.configuration.BotConfig;
-import edu.java.configuration.GitHubConfig;
-import edu.java.configuration.StackOverflowConfig;
+import edu.java.configuration.properties.ApplicationConfig;
+import edu.java.configuration.properties.BotConfig;
+import edu.java.configuration.properties.GitHubConfig;
+import edu.java.configuration.properties.StackOverflowConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.ConfigurationPropertiesScan;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
 @EnableScheduling
-@EnableConfigurationProperties({
-        ApplicationConfig.class,
-        GitHubConfig.class,
-        StackOverflowConfig.class,
-        BotConfig.class
-})
+@ConfigurationPropertiesScan
 public class ScrapperApplication {
     public static void main(String[] args) {
         SpringApplication.run(ScrapperApplication.class, args);
