+++ scrapper/src/main/java/edu/java/configuration/BotClientConfiguration.java
@@ -1,6 +1,7 @@
 package edu.java.configuration;
 
 import edu.java.client.bot.BotClient;
+import edu.java.configuration.properties.BotConfig;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
