+++ scrapper/src/main/java/edu/java/configuration/properties/BotConfig.java
@@ -1,12 +1,12 @@
-package edu.java.bot.configuration;
+package edu.java.configuration.properties;
 
 import jakarta.validation.constraints.NotEmpty;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.validation.annotation.Validated;
 
 @Validated
-@ConfigurationProperties(prefix = "scrapper-client", ignoreUnknownFields = false)
-public record ScrapperConfig(
+@ConfigurationProperties(prefix = "bot-client", ignoreUnknownFields = false)
+public record BotConfig(
         @NotEmpty
         String baseUrl
 ) {
