+++ scrapper/src/main/java/edu/java/configuration/properties/StackOverflowConfig.java
@@ -1,4 +1,4 @@
-package edu.java.configuration;
+package edu.java.configuration.properties;
 
 import jakarta.validation.constraints.NotNull;
 import org.springframework.boot.context.properties.ConfigurationProperties;
